// src/utils/interfaces/revenue.ts
export interface IRevenue {
    description: string;
    value: number;
    release: string;
    dueDate: string;
    effectiveDate: string;
    recurrence: string;
    installment: number;
    installments: number;
    category: string;
    subcategory: string;
    account: string;
    notes: string;
    hash: string;
}